Add test helper to get result from async functions
